Add files via upload
Button now has three modes, gain control, buffer clear, freeze. Hold down the button for 10 seconds, when the LEDs come on press again within one second to confirm mode change. This is to avoid unwanted mode changes when using cv control. If in gain control mode not confirming will lock mod depth to 3.125%. In buffer clear mode, a rising edge will trigger the buffers to clear. In freeze mode, holding the button will freeze the buffers.

Tried to reduce clicks as much as possible when clearing the buffers. Still not as clean as with the software. Need help fixing this. There's probably a more elegant solution.